Allow the user to supply a custom ssl context
The new ImapTransport parameter ssl_context replaces the usesslcontext
parameter and allows the user to supply their own ssl context object. If
ssl_context is not given, but ssl is true, python's default ssl context
is used. That default context is the one that actually does some
certificate checks, such as whether the hostname matches the names given
in the server's certificate and not the default context used by
IMAP4_SSL when instantiated with ssl_context=None which does not
certificate checks at all.

The Imbox class is extended with the same ssl_context parameter which is
simply passed through to ImapTransport.

This commit together with the previous commits from Dustin Demuth
changes Imbox in a slightly incompatible way: SSL-Certificates are now
checked by default whereas before they were not checked at all. This
improves security substantially but users need to be aware that working
programs might start raising exceptions due to failing certificate
checks.